feat: allow empty SuperclusterBase
Previously, attempting to init an empty SuperclusterBase instance would
raise an EmptySuperclusterError. This caused complexity when using
SuperclusterBase.copy to exclude clusters or members, because an empty
SuperclusterBase is a good stopping point (not an error) for some
algorithms. Now, you can init a SuperclusterBase, but attempting to get
or set a quality condition will raise an EmptySuperclusterError.

BREAKING CHANGE: Any code that relied on catching EmptySuperclusterError
as a stopping condition will now fail.